pyproject.toml
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "django-template"
version = "0.1.0"
[tool.poetry.dependencies]
Django = "5.0.2"
python = "^3.10"
streamlit = "^1.49.1"
psycopg2-binary = "^2.9.10"
python-docx = "^1.2.0"
reportlab = "^4.4.4"
stripe = "^12.5.1"
pypdf2 = "^3.0.1"
python-dotenv = "^1.1.1"
pillow = "^11.3.0"
requests = "^2.32.5"
openai = "^1.108.1"
djangorestframework = "3.14.0"
django-cors-headers = "4.3.1"
django-filter = "23.5"
[tool.poetry.dev-dependencies]

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"